requirements.txt
pypdf
pymupdf
faiss-cpu
sentence-transformers
openai
python-dotenv
flask
flask-cors
waitress
langchain
langchain-openai
langchain-community
langchain-text-splitters